+++ Assets/UGF.Application.Runtime.Tests/TestApplicationLauncher.cs
@@ -1,8 +1,11 @@
+using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
+using Object = UnityEngine.Object;
 
 namespace UGF.Application.Runtime.Tests
 {
@@ -21,15 +24,7 @@ namespace UGF.Application.Runtime.Tests
                 Assert.False(HasApplication);
             }
 
-            protected override async Task PreloadResourcesAsync()
-            {
-                await Task.Yield();
-
-                Assert.True(IsLaunched);
-                Assert.False(HasApplication);
-            }
-
-            protected override IApplication CreateApplication()
+            protected override IApplication CreateApplication(IApplicationResources resources)
             {
                 Assert.True(IsLaunched);
                 Assert.False(HasApplication);
@@ -55,16 +50,6 @@ namespace UGF.Application.Runtime.Tests
                 Assert.False(application.GetModule<ModuleAsync>().IsInit);
             }
 
-            protected override async Task InitializeModulesAsync(IApplication application)
-            {
-                await base.InitializeModulesAsync(application);
-
-                Assert.True(IsLaunched);
-                Assert.False(HasApplication);
-                Assert.True(application.GetModule<Module>().IsInit);
-                Assert.True(application.GetModule<ModuleAsync>().IsInit);
-            }
-
             protected override void OnLaunched(IApplication application)
             {
                 base.OnLaunched(application);
@@ -75,24 +60,14 @@ namespace UGF.Application.Runtime.Tests
                 Assert.True(application.GetModule<ModuleAsync>().IsInit);
             }
 
-            protected override void OnStop(IApplication application)
+            protected override void OnStopped(IApplication application)
             {
-                base.OnStop(application);
+                base.OnStopped(application);
 
                 Assert.False(IsLaunched);
                 Assert.True(HasApplication);
                 Assert.True(application.GetModule<Module>().IsInit);
                 Assert.True(application.GetModule<ModuleAsync>().IsInit);
-            }
-
-            protected override void OnStopped()
-            {
-                base.OnStopped();
-
-                Assert.False(IsLaunched);
-                Assert.False(HasApplication);
-                Assert.False(m_module.IsInit);
-                Assert.False(m_moduleAsync.IsInit);
 
                 m_module = null;
                 m_moduleAsync = null;
@@ -106,6 +81,16 @@ namespace UGF.Application.Runtime.Tests
 
         private class Application : ApplicationBase
         {
+            protected override async Task OnInitializeAsync()
+            {
+                foreach (KeyValuePair<Type, IApplicationModule> pair in Modules)
+                {
+                    if (pair.Value is IApplicationModuleAsync module)
+                    {
+                        await module.InitializeAsync();
+                    }
+                }
+            }
         }
 
         private class Module : ApplicationModuleBase
@@ -127,11 +112,11 @@ namespace UGF.Application.Runtime.Tests
             }
         }
 
-        private class ModuleAsync : ApplicationModuleBaseAsync
+        private class ModuleAsync : ApplicationModuleBase, IApplicationModuleAsync
         {
             public bool IsInit { get; private set; }
 
-            public override async Task InitializeAsync()
+            public async Task InitializeAsync()
             {
                 await Task.Yield();