+++ Assets/UGF.Application.Runtime.Tests/TestKernelApplication.cs
@@ -57,12 +57,12 @@ namespace UGF.Application.Runtime.Tests
         private class ModuleInfo : IApplicationModuleInfo
         {
             public Type RegisterType { get; }
-            public IApplicationModuleBuilder Builder { get; }
+            public ApplicationModuleBuildHandler Builder { get; }
 
             public ModuleInfo(IApplicationModule module)
             {
                 RegisterType = module.GetType();
-                Builder = new ApplicationModuleBuilder(application => module);
+                Builder = application => module;
             }
         }
 
@@ -78,7 +78,8 @@ namespace UGF.Application.Runtime.Tests
                 }
             };
 
-            var application = new ApplicationConfigured(config, false);
+            var resources = new ApplicationResources { config };
+            var application = new ApplicationConfigured(resources, false);
 
             application.Initialize();
 
@@ -105,7 +106,8 @@ namespace UGF.Application.Runtime.Tests
                 }
             };
 
-            var application = new ApplicationConfigured(config, false);
+            var resources = new ApplicationResources { config };
+            var application = new ApplicationConfigured(resources, false);
 
             application.Initialize();
 
@@ -128,7 +130,8 @@ namespace UGF.Application.Runtime.Tests
                 }
             };
 
-            var application = new ApplicationConfigured(config, false);
+            var resources = new ApplicationResources { config };
+            var application = new ApplicationConfigured(resources, false);
 
             application.AddModule(new ModuleC(() => order.Add("moduleC")));
             application.Initialize();
@@ -153,7 +156,8 @@ namespace UGF.Application.Runtime.Tests
                 }
             };
 
-            var application = new ApplicationConfigured(config, false);
+            var resources = new ApplicationResources { config };
+            var application = new ApplicationConfigured(resources, false);
 
             application.Initialize();
             application.Uninitialize();
@@ -177,7 +181,8 @@ namespace UGF.Application.Runtime.Tests
                 }
             };
 
-            var application = new ApplicationConfigured(config, false);
+            var resources = new ApplicationResources { config };
+            var application = new ApplicationConfigured(resources, false);
 
             application.Initialize();