+++ Packages/UGF.Application/Runtime/ApplicationConfigAssetBase.cs
@@ -1,9 +1,12 @@
-using UnityEngine;
-
 namespace UGF.Application.Runtime
 {
-    public abstract class ApplicationConfigAssetBase : ScriptableObject
+    public abstract class ApplicationConfigAssetBase : ApplicationResourceAsset
     {
         public abstract IApplicationConfig GetConfig();
+
+        public override object GetResource()
+        {
+            return GetConfig();
+        }
     }
 }