+++ Packages/UGF.Application/Runtime/ApplicationResourceAsset.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace UGF.Application.Runtime
+{
+    public abstract class ApplicationResourceAsset : ScriptableObject
+    {
+        public abstract object GetResource();
+    }
+}