+++ Packages/UGF.Application/Runtime/ApplicationUnity.cs
@@ -15,8 +15,9 @@ namespace UGF.Application.Runtime
         /// <summary>
         /// Creates application with specified arguments.
         /// </summary>
+        /// <param name="resources">The application resources.</param>
         /// <param name="provideStaticInstance">The value that determines whether to provide static instance via <see cref="ApplicationInstance"/>.</param>
-        public ApplicationUnity(bool provideStaticInstance)
+        public ApplicationUnity(IApplicationResources resources = null, bool provideStaticInstance = false) : base(resources)
         {
             ProvideStaticInstance = provideStaticInstance;
         }