+++ Packages/UGF.Application/Runtime/IApplicationLauncherEventHandler.cs
@@ -1,0 +1,9 @@
+﻿namespace UGF.Application.Runtime
+{
+    public interface IApplicationLauncherEventHandler
+    {
+        void OnLaunched(IApplication application);
+        void OnStopped(IApplication application);
+        void OnQuitting(IApplication application);
+    }
+}