+++ Packages/UGF.Application/Runtime/IApplicationModuleAsync.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace UGF.Application.Runtime
+{
+    public interface IApplicationModuleAsync
+    {
+        Task InitializeAsync();
+    }
+}